Improving the activity locations section.

+++ trunk/ccafspr_ip/src/main/java/org/cgiar/ccafs/ap/data/manager/impl/LocationManagerImpl.java
@@ -347,6 +347,11 @@ public class LocationManagerImpl implements LocationManager {
 
     for (Location location : activityLocations) {
       if (location.isOtherLocation()) {
+        OtherLocation otherLocation = (OtherLocation) location;
+        if (otherLocation.getType().getId() == APConstants.LOCATION_TYPE_CCAFS_SITE) {
+          continue;
+        }
+
         removed = locationDAO.removeLocation(location.getId());
         allRemoved = (removed && allRemoved);
       }
